editor/requirements.txt
astroid==3.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
dill==0.3.8
dnspython==2.6.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
isort==5.13.2
jedi==0.19.1
mccabe==0.7.0
moviepy==1.0.3
numpy==1.26.4
parso==0.8.3
pika==1.3.2
pilgram==1.2.1
pilgram2==2.0.4
pillow==10.2.0
platformdirs==4.2.0
proglog==0.1.10
pylint==3.0.3
pymongo==4.6.1
requests==2.31.0
setuptools==69.1.0
tomlkit==0.12.3
tqdm==4.66.2
urllib3==2.2.1
